SOM network works

+++ SOM-BIAI_project/include/SubframeCompressed.h
@@ -10,26 +10,26 @@ class SubframeCompressed {
     int lumaWinnerIndex;
     int redChromaWinnerIndex;
     int blueChromaWinnerIndex;
-    uint8_t lumaValue;
-    uint8_t redChromaValue;
-    uint8_t blueChromaValue;
+    double lumaValue;
+    double redChromaValue;
+    double blueChromaValue;
 
    public:
     SubframeCompressed();
     SubframeCompressed(int lumaWinnerIndex, int redChromaWinnerIndex,
-                       int blueChromaWinnerIndex, uint8_t lumaValue,
-                       uint8_t redChromaValue, uint8_t blueChromaValue);
-    void setLumaValue(uint8_t lumaValue);
-    uint8_t getLumaValue();
-    void setRedChromaValue(uint8_t redChromaValue);
-    uint8_t getRedChromaValue();
-    void setBlueChromaValue(uint8_t blueChromaValue);
-    uint8_t getBlueChromaValue();
-    void setLumaWinnerIndex(uint8_t lumaWinnerIndex);
-    uint8_t getLumaWinnerIndex();
-    void setRedChromaWinnerIndex(uint8_t redChromaWinnerIndex);
-    uint8_t getRedChromaWinnerIndex();
-    void setBlueChromaWinnerIndex(uint8_t blueChromaWinnerIndex);
-    uint8_t getBlueChromaWinnerIndex();
+                       int blueChromaWinnerIndex, double lumaValue,
+                       double redChromaValue, double blueChromaValue);
+    void setLumaValue(double lumaValue);
+    double getLumaValue();
+    void setRedChromaValue(double redChromaValue);
+    double getRedChromaValue();
+    void setBlueChromaValue(double blueChromaValue);
+    double getBlueChromaValue();
+    void setLumaWinnerIndex(double lumaWinnerIndex);
+    double getLumaWinnerIndex();
+    void setRedChromaWinnerIndex(double redChromaWinnerIndex);
+    double getRedChromaWinnerIndex();
+    void setBlueChromaWinnerIndex(double blueChromaWinnerIndex);
+    double getBlueChromaWinnerIndex();
 };
 }  // namespace SOM
